Smooth transition in the graph
When I was visualizing the work of various filters, the problem arose that it is not always visible how the filters work. Therefore, I made a similar algorithm that approximates the chart zone I need . At the moment, the transformation is based on a linear function, since this is quite enough in the presentation.