Odometry for EV3 robot
Unfortunately, I only have an EV3 robot for testing algorithms. But the block itself works quite slowly, so it is possible to dynamically find the robot's position in space, but with great difficulty. The algorithm uses values of small increments of encoder values from motors. So I just wrote down the values of these increments and analyzed them separately in Matlab. As expected, different iteration times strongly affect accuracy. But regardless of the selected iteration time , the accuracy still sucks.